schemas, getContacts, addContact work
