+++ controllers/contacts-controllers.js
@@ -1,50 +1,52 @@
-const contacts = require("../models/contacts");
+// const contacts = require("../models/contacts");
+
+const {Contact} = require("../models/contact")
 
 const { ctrlWrapper } = require("../utils");
 
 const { HttpError } = require("../helpers");
 
 const getAllContacts = async (req, res) => {
-  const result = await contacts.listContacts();
+  const result = await Contact.find();
   res.json(result);
 };
 
-const getContactById = async (req, res) => {
-  const { id } = req.params;
-  const result = await contacts.getContactById(id);
-  if (!result) {
-    throw HttpError(404, `Not found`);
-  }
-  res.json(result);
-};
+// const getContactById = async (req, res) => {
+//   const { id } = req.params;
+//   const result = await contacts.getContactById(id);
+//   if (!result) {
+//     throw HttpError(404, `Not found`);
+//   }
+//   res.json(result);
+// };
 
 const addNewContact = async (req, res) => {
-  const result = await contacts.addContact(req.body);
+  const result = await Contact.create(req.body);
   res.status(201).json(result);
 };
 
-const deleteContactById = async (req, res) => {
-  const { id } = req.params;
-  const result = await contacts.removeContact(id);
-  if (!result) {
-    throw HttpError(404, `Not found`);
-  }
-  res.json({ message: "contact deleted" });
-};
+// const deleteContactById = async (req, res) => {
+//   const { id } = req.params;
+//   const result = await contacts.removeContact(id);
+//   if (!result) {
+//     throw HttpError(404, `Not found`);
+//   }
+//   res.json({ message: "contact deleted" });
+// };
 
-const updateOneContact = async (req, res) => {
-  const { id } = req.params;
-  const result = await contacts.updateContact(id, req.body);
-  if (!result) {
-    throw HttpError(404, `Not found`);
-  }
-  res.json(result);
-};
+// const updateOneContact = async (req, res) => {
+//   const { id } = req.params;
+//   const result = await contacts.updateContact(id, req.body);
+//   if (!result) {
+//     throw HttpError(404, `Not found`);
+//   }
+//   res.json(result);
+// };
 
 module.exports = {
   getAllContacts: ctrlWrapper(getAllContacts),
-  getContactById: ctrlWrapper(getContactById),
+//   getContactById: ctrlWrapper(getContactById),
   addNewContact: ctrlWrapper(addNewContact),
-  deleteContactById: ctrlWrapper(deleteContactById),
-  updateOneContact: ctrlWrapper(updateOneContact),
+//   deleteContactById: ctrlWrapper(deleteContactById),
+//   updateOneContact: ctrlWrapper(updateOneContact),
 };
