+++ routes/api/contacts-routes.js
@@ -4,18 +4,18 @@ const ctrl = require("../../controllers/contacts-controllers");
 
 const { validateBody } = require("../../utils");
 
-const schema = require("../../schemas/contacts-schema");
+const {schemas} = require("../../models/contact");
 
 const router = express.Router();
 
 router.get("/", ctrl.getAllContacts);
 
-router.get("/:contactId", ctrl.getContactById);
+// router.get("/:contactId", ctrl.getContactById);
 
-router.post("/", validateBody(schema.addSchema), ctrl.addNewContact);
+router.post("/", validateBody(schemas.addSchema), ctrl.addNewContact);
 
-router.delete("/:contactId", ctrl.deleteContactById);
+// router.delete("/:contactId", ctrl.deleteContactById);
 
-router.put("/:contactId", validateBody(schema.addSchema), ctrl.updateOneContact);
+// router.put("/:contactId", validateBody(schema.addSchema), ctrl.updateOneContact);
 
 module.exports = router;
