+++ schemas/contacts-schema.js
@@ -1,11 +1,11 @@
-const Joi = require("joi");
+// const Joi = require("joi");
 
-const addSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().required(),
-    phone: Joi.string().required(),
-  });
+// const addSchema = Joi.object({
+//     name: Joi.string().required(),
+//     email: Joi.string().required(),
+//     phone: Joi.string().required(),
+//   });
 
-module.exports = {
-    addSchema,
-}
+// module.exports = {
+//     addSchema,
+// }
